+++ frontend/src/app/chattings/page.tsx
@@ -1,4 +1,4 @@
-import Chattings from "../../components/pages/Chattings";
+import Chattings from "./_components/Chattings";
 
 export default function ChattingsPage() {
   return <Chattings />;
