+++ frontend/src/app/login/page.tsx
@@ -1,4 +1,4 @@
-import Login from "../../components/pages/Login";
+import Login from "./_components/Login";
 
 export default function LoginPage() {
   return <Login />;
